requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-generativeai==0.3.2
requests==2.31.0
beautifulsoup4==4.12.2
playwright==1.40.0
pydantic==2.5.0
pandas==2.1.4
supabase==2.3.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
